Moved Client class to a separate file.
